deprecated Student constructor specifying an ID
Also added a new constructor and some static / const variables to use
instead. I think this way of doing things is better, although I kept
your style intact.. we really should have an enforced consistent style
for all the code though. I don't care for this project but it's
something to think about in the future.